feat(frontend): improve and fix error homepage styles
