+++ src/css/custom.css
@@ -2,8 +2,8 @@
  * You can override the default Infima variables here.
  * Note: this is not a complete list of --ifm- variables.
  */
-@import url("https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;700&family=PT+Sans:wght@400;700&display=swap");
-:root {
+ @import url("https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap");
+ :root {
   --ifm-color-primary: #3ebbd3;
   --ifm-color-primary-dark: #212121;
   --ifm-color-primary-darker: #424242;
@@ -13,7 +13,7 @@
   --ifm-color-primary-lightest: #b6ebf3;
 }
 body {
-  font-family: "PT Sans", sans-serif;
+  font-family: "Roboto", sans-serif;
 }
 img {
   aspect-ratio: attr(width) / attr(height);
@@ -24,8 +24,7 @@ img {
   margin-right: auto;
 }
 h1 {
-  font-family: "PT Sans", sans-serif;
-  /*font-size: 101.4px;*/
+  font-family: "Roboto", sans-serif;
   font-weight: normal;
   font-stretch: normal;
   font-style: normal;
@@ -33,8 +32,7 @@ h1 {
   letter-spacing: -1.5px;
 }
 h2 {
-  font-family: "PT Sans", sans-serif;
-  /*font-size: 63.4px;*/
+  font-family: "Roboto", sans-serif;
   font-weight: normal;
   font-stretch: normal;
   font-style: normal;
@@ -42,8 +40,7 @@ h2 {
   letter-spacing: -0.5px;
 }
 h3 {
-  font-family: "PT Sans", sans-serif;
-  /*font-size: 50.7px;*/
+  font-family: "Roboto", sans-serif;
   font-weight: normal;
   font-stretch: normal;
   font-style: normal;
@@ -51,8 +48,7 @@ h3 {
   letter-spacing: normal;
 }
 h4 {
-  font-family: "PT Sans", sans-serif;
-  /*font-size: 35.9px;*/
+  font-family: "Roboto", sans-serif;
   font-weight: normal;
   font-stretch: normal;
   font-style: normal;
@@ -60,8 +56,7 @@ h4 {
   letter-spacing: 0.25px;
 }
 h5 {
-  font-family: "PT Sans", sans-serif;
-  /*font-size: 25.4px;*/
+  font-family: "Roboto", sans-serif;
   font-weight: normal;
   font-stretch: normal;
   font-style: normal;
@@ -69,8 +64,7 @@ h5 {
   letter-spacing: normal;
 }
 h6 {
-  font-family: "PT Sans", sans-serif;
-  /*font-size: 21.1px;*/
+  font-family: "Roboto", sans-serif;
   font-weight: bold;
   font-stretch: normal;
   font-style: normal;
@@ -78,8 +72,7 @@ h6 {
   letter-spacing: 0.25px;
 }
 button {
-  font-family: "PT Sans", sans-serif;
-  /*font-size: 14.8px;*/
+  font-family: "Roboto", sans-serif;
   font-weight: normal;
   font-stretch: normal;
   font-style: normal;
@@ -87,7 +80,7 @@ button {
   letter-spacing: 1.25px;
 }
 p {
-  font-family: "Open Sans", sans-serif;
+  font-family: "Roboto", sans-serif;
   font-size: 15.8px;
   font-weight: normal;
   font-stretch: normal;
